fix: moved configs dynamicModifiers files to separate file
